rag-document-parser-app-1/requirements.txt
numpy
pandas
PyPDF2
scikit-learn
pinecone-client
openai
fastapi
uvicorn
pydantic
pytest
loguru